Updated module augmentation for ckfinder package.
