+++ packages/ckeditor5-build-inline/src/ckeditor.ts
@@ -13,7 +13,7 @@ import Bold from '@ckeditor/ckeditor5-basic-styles/src/bold';
 import Italic from '@ckeditor/ckeditor5-basic-styles/src/italic';
 import BlockQuote from '@ckeditor/ckeditor5-block-quote/src/blockquote';
 import { CKBox } from '@ckeditor/ckeditor5-ckbox';
-import CKFinder from '@ckeditor/ckeditor5-ckfinder/src/ckfinder';
+import { CKFinder } from '@ckeditor/ckeditor5-ckfinder';
 import EasyImage from '@ckeditor/ckeditor5-easy-image/src/easyimage';
 import Heading from '@ckeditor/ckeditor5-heading/src/heading';
 import Image from '@ckeditor/ckeditor5-image/src/image';
